workflow

+++ src/Core/LunchLine.Application/DTOs/EmployeeDTOs/EmployeeCreateDTO.cs
@@ -8,5 +8,5 @@ public record EmployeeCreateDTO
     public string PhoneNumber { get; set; }
     public string PositionId { get; set; }
     public bool? Gender { get; set; }
-    public float Salary { get; set; }
+    public decimal Salary { get; set; }
 }
